+++ src/main/java/jcl/reader/function/macrofunction/RightParenthesisReaderMacroFunction.java
@@ -1,9 +1,11 @@
 package jcl.reader.function.macrofunction;
 
 import jcl.LispStruct;
-import jcl.structs.conditions.exceptions.ReaderErrorException;
 import jcl.reader.impl.Reader;
+import jcl.structs.conditions.exceptions.ReaderErrorException;
 import jcl.syntax.CharacterConstants;
+
+import java.math.BigInteger;
 
 /**
  * Implements the ')' Lisp reader macro.
@@ -11,7 +13,7 @@ import jcl.syntax.CharacterConstants;
 public class RightParenthesisReaderMacroFunction extends ReaderMacroFunction {
 
 	@Override
-	public LispStruct readMacro(final int codePoint, final Reader reader, final Integer numArg) {
+	public LispStruct readMacro(final int codePoint, final Reader reader, final BigInteger numArg) {
 		assert codePoint == CharacterConstants.RIGHT_PARENTHESIS;
 		throw new ReaderErrorException("Unmatched close parenthesis.");
 	}
