+++ src/main/java/jcl/reader/function/macrofunction/SharpIllegalReaderMacroFunction.java
@@ -4,13 +4,15 @@ import jcl.LispStruct;
 import jcl.structs.conditions.exceptions.ReaderErrorException;
 import jcl.reader.impl.Reader;
 
+import java.math.BigInteger;
+
 /**
  * Implements the illegal '#??" Lisp reader macros.
  */
 public class SharpIllegalReaderMacroFunction extends ReaderMacroFunction {
 
 	@Override
-	public LispStruct readMacro(final int codePoint, final Reader reader, final Integer numArg) {
+	public LispStruct readMacro(final int codePoint, final Reader reader, final BigInteger numArg) {
 		throw new ReaderErrorException("Illegal sharp character " + codePoint);
 	}
 }
