+++ src/main/java/jcl/reader/function/macrofunction/SharpOReaderMacroFunction.java
@@ -5,16 +5,18 @@ import jcl.reader.function.RationalReader;
 import jcl.reader.impl.Reader;
 import jcl.syntax.CharacterConstants;
 
+import java.math.BigInteger;
+
 /**
  * Implements the '#o' Lisp reader macro.
  */
 public class SharpOReaderMacroFunction extends ReaderMacroFunction {
 
 	@Override
-	public LispStruct readMacro(final int codePoint, final Reader reader, final Integer numArg) {
+	public LispStruct readMacro(final int codePoint, final Reader reader, final BigInteger numArg) {
 		assert (codePoint == CharacterConstants.LATIN_SMALL_LETTER_O) || (codePoint == CharacterConstants.LATIN_CAPITAL_LETTER_O);
 
 		final RationalReader macroFunctionReader = new RationalReader(reader);
-		return macroFunctionReader.readRationalToken(8);
+		return macroFunctionReader.readRationalToken(BigInteger.valueOf(8));
 	}
 }
