+++ src/main/java/jcl/reader/function/macrofunction/SharpRReaderMacroFunction.java
@@ -5,13 +5,15 @@ import jcl.reader.function.RationalReader;
 import jcl.reader.impl.Reader;
 import jcl.syntax.CharacterConstants;
 
+import java.math.BigInteger;
+
 /**
  * Implements the '#r' Lisp reader macro.
  */
 public class SharpRReaderMacroFunction extends ReaderMacroFunction {
 
 	@Override
-	public LispStruct readMacro(final int codePoint, final Reader reader, final Integer numArg) {
+	public LispStruct readMacro(final int codePoint, final Reader reader, final BigInteger numArg) {
 		assert (codePoint == CharacterConstants.LATIN_SMALL_LETTER_R) || (codePoint == CharacterConstants.LATIN_CAPITAL_LETTER_R);
 
 		final RationalReader macroFunctionReader = new RationalReader(reader);
